+++ app/src/main/AndroidManifest.xml
@@ -15,28 +15,28 @@
         <activity android:name="admin.AdminCheckNewProductsActivity"></activity>
         <activity android:name="admin.AdminHomeActivity" />
         <activity
-            android:name="Sellers.SellerHomeActivity"
+            android:name="sellers.SellerHomeActivity"
             android:label="@string/title_activity_seller_home" />
-        <activity android:name="Sellers.SellerLoginActivity" />
-        <activity android:name="Sellers.SellerRegistrationActivity" />
-        <activity android:name="Buyers.ResetPasswordActivity" />
+        <activity android:name="sellers.SellerLoginActivity" />
+        <activity android:name="sellers.SellerRegistrationActivity" />
+        <activity android:name="buyers.ResetPasswordActivity" />
         <activity android:name="admin.AdminMaintainProductsActivity" />
-        <activity android:name="Buyers.SearchProductsActivity" />
+        <activity android:name="buyers.SearchProductsActivity" />
         <activity android:name="admin.AdminUserProductsActivity" />
         <activity android:name="admin.AdminNewOrdersActivity" />
-        <activity android:name="Buyers.ConfirmFinalOrderActivity" />
-        <activity android:name="Buyers.CartActivity" />
+        <activity android:name="buyers.ConfirmFinalOrderActivity" />
+        <activity android:name="buyers.CartActivity" />
         <activity
-            android:name="Buyers.HomeActivity"
+            android:name="buyers.HomeActivity"
             android:label="@string/title_activity_home"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name="Sellers.SellerProductCategoryActivity" />
-        <activity android:name="Sellers.SellerAddNewProductActivity" />
-        <activity android:name="Buyers.RegisterActivity" />
-        <activity android:name="Buyers.LoginActivity" />
-        <activity android:name="Buyers.SettinsActivity" />
-        <activity android:name="Buyers.ProductDetailsActivity" />
-        <activity android:name="Buyers.MainActivity">
+        <activity android:name="sellers.SellerProductCategoryActivity" />
+        <activity android:name="sellers.SellerAddNewProductActivity" />
+        <activity android:name="buyers.RegisterActivity" />
+        <activity android:name="buyers.LoginActivity" />
+        <activity android:name="buyers.SettinsActivity" />
+        <activity android:name="buyers.ProductDetailsActivity" />
+        <activity android:name="buyers.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
