+++ app/src/main/java/sellers/SellerLoginActivity.java
@@ -1,4 +1,4 @@
-package Sellers;
+package sellers;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
