+++ app/src/main/java/sellers/SellerProductCategoryActivity.java
@@ -1,4 +1,4 @@
-package Sellers;
+package sellers;
 
 import androidx.appcompat.app.AppCompatActivity;
