+++ app/src/main/res/layout/activity_cart.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Buyers.CartActivity">
+    tools:context="buyers.CartActivity">
 
     <RelativeLayout
         android:id="@+id/rl11"
