+++ app/src/main/res/layout/activity_confirm_final_order.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Buyers.ConfirmFinalOrderActivity">
+    tools:context="buyers.ConfirmFinalOrderActivity">
 
     <TextView
         android:id="@+id/txt"
