+++ app/src/main/res/layout/activity_main.xml
@@ -5,30 +5,24 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/welcome4"
-    tools:context="Buyers.MainActivity">
-
-    <ImageView
-        android:id="@+id/app_logo"
-        android:layout_width="300dp"
-        android:layout_height="100dp"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="60dp"
-        />
+    android:background="@drawable/rr"
+    tools:context="buyers.MainActivity">
 
 
-    
+
+
+
     <Button
         android:id="@+id/main_login_btn"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_above="@id/seller_begin"
         android:layout_alignParentStart="true"
-        android:layout_marginBottom="15dp"
+        android:layout_marginBottom="10dp"
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
         android:background="@drawable/buttons"
-        android:padding="15dp"
+        android:padding="20dp"
         android:textAllCaps="false"
         android:textSize="18sp"
         android:text="Already have an account? Login"
@@ -41,11 +35,11 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_above="@+id/main_login_btn"
-        android:layout_marginBottom="17dp"
+        android:layout_marginBottom="5dp"
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
         android:background="@drawable/input_design"
-        android:padding="15dp"
+        android:padding="20dp"
         android:textAllCaps="false"
         android:textSize="18sp"
         android:text="Join Now"
