+++ app/src/main/res/layout/activity_product_details.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="Buyers.ProductDetailsActivity">
+    tools:context="buyers.ProductDetailsActivity">
 
 
     <ImageView
