+++ app/src/main/res/layout/activity_register.xml
@@ -4,38 +4,27 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/second1"
-    tools:context="Buyers.RegisterActivity">
-
-
-    <ImageView
-        android:id="@+id/register_applogo"
-        android:layout_width="300dp"
-        android:layout_height="100dp"
-
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="190dp"
-        />
+    android:background="@drawable/ee"
+    tools:context="buyers.RegisterActivity">
 
 
     <EditText
         android:id="@+id/register_username_input"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@+id/register_applogo"
-        android:background="@drawable/input_design"
-        android:padding="20dp"
+        android:layout_alignParentTop="true"
         android:layout_marginLeft="15dp"
+        android:layout_marginTop="300dp"
         android:layout_marginRight="15dp"
+        android:background="@drawable/input_design"
         android:hint="Name"
         android:inputType="text"
-        android:textColor="@color/colorPrimaryDark"
-        android:textColorHint="@color/colorPrimaryDark"
+        android:padding="20dp"
+        android:textColor="@android:color/black"
+        android:textColorHint="@android:color/black"
         android:textSize="17sp"
-        android:textStyle="bold"
-        android:layout_marginTop="110dp"
 
-        />
+        android:textStyle="bold" />
 
 
 
@@ -50,8 +39,8 @@
         android:layout_marginRight="15dp"
         android:hint="Phone Number"
         android:inputType="number"
-        android:textColor="@color/colorPrimaryDark"
-        android:textColorHint="@color/colorPrimaryDark"
+        android:textColor="@android:color/black"
+        android:textColorHint="@android:color/black"
         android:textSize="17sp"
         android:textStyle="bold"
         android:layout_marginTop="6dp"
@@ -70,8 +59,8 @@
         android:layout_marginTop="6dp"
         android:hint="Password"
         android:inputType="textPassword"
-        android:textColor="@color/colorPrimaryDark"
-        android:textColorHint="@color/colorPrimaryDark"
+        android:textColor="@android:color/black"
+        android:textColorHint="@android:color/black"
         android:textSize="17sp"
         android:textStyle="bold"
