+++ app/src/main/res/layout/activity_reset_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Buyers.ResetPasswordActivity">
+    tools:context="buyers.ResetPasswordActivity">
 
 
     <RelativeLayout
