+++ app/src/main/res/layout/activity_search_products.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Buyers.SearchProductsActivity">
+    tools:context="buyers.SearchProductsActivity">
 
     <RelativeLayout
         android:id="@+id/r5"
