+++ app/src/main/res/layout/activity_seller_add_new_product.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Sellers.SellerAddNewProductActivity">
+    tools:context="sellers.SellerAddNewProductActivity">
     
     
     <ImageView
