+++ app/src/main/res/layout/activity_seller_home.xml
@@ -6,7 +6,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Sellers.SellerHomeActivity">
+    tools:context="sellers.SellerHomeActivity">
 
     <RelativeLayout
         android:id="@+id/rel1"
