+++ app/src/main/res/layout/activity_seller_login.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Sellers.SellerLoginActivity">
+    tools:context="sellers.SellerLoginActivity">
 
 
     <ImageView
