+++ app/src/main/res/layout/activity_seller_registration.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Sellers.SellerRegistrationActivity">
+    tools:context="sellers.SellerRegistrationActivity">
 
 
     <ImageView
@@ -22,7 +22,7 @@
         android:text="Seller Registration Form"
         android:gravity="center"
         android:layout_below="@+id/img_seller"
-        android:layout_marginTop="30dp"
+        android:layout_marginTop="5dp"
         android:textStyle="bold"
         android:textSize="24dp"
         android:textColor="@android:color/holo_orange_dark"
@@ -37,7 +37,7 @@
         android:hint="Name..."
         android:inputType="text"
         android:textSize="16dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         />
     <EditText
         android:id="@+id/seller_phone"
@@ -49,7 +49,7 @@
         android:hint="Phone..."
         android:inputType="number"
         android:textSize="16dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         />
     <EditText
         android:id="@+id/seller_email"
@@ -61,7 +61,7 @@
         android:hint="Email..."
         android:inputType="textEmailAddress"
         android:textSize="16dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         />
     <EditText
         android:id="@+id/seller_password"
@@ -73,7 +73,7 @@
         android:hint="Password..."
         android:inputType="text"
         android:textSize="16dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         />
     <EditText
         android:id="@+id/seller_address"
@@ -85,7 +85,7 @@
         android:hint="Shop - Business Address..."
         android:inputType="text"
         android:textSize="16dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         />
     <Button
         android:id="@+id/seller_register_btn"
@@ -93,7 +93,7 @@
         android:layout_height="wrap_content"
         android:layout_marginLeft="15dp"
         android:layout_marginRight="15dp"
-        android:layout_marginTop="10dp"
+        android:layout_marginTop="8dp"
         android:layout_below="@id/seller_address"
         android:background="@android:color/holo_orange_dark"
         android:text="Register"
@@ -114,7 +114,7 @@
         android:textSize="18dp"
         android:layout_alignParentBottom="true"
         android:textColor="@android:color/holo_orange_light"
-        android:layout_marginBottom="15dp"
+        android:layout_marginBottom="5dp"
         />
