+++ app/src/main/res/layout/activity_settins.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Buyers.SettinsActivity">
+    tools:context="buyers.SettinsActivity">
 
 
     <com.google.android.material.appbar.AppBarLayout
