+++ app/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Buyers.HomeActivity">
+    tools:context="buyers.HomeActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_height="wrap_content"
