+++ app/src/main/res/layout/fragment_notifications.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="Sellers.ui.notifications.NotificationsFragment">
+    tools:context="sellers.ui.notifications.NotificationsFragment">
 
     <TextView
         android:id="@+id/text_notifications"
